+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Localized the new item and actor drop-down menus so that they appear more professional.
 - Critical Injuries are now items so that injuries can be premade, dragged to the character sheet, and used more easily in critical injury tables.
   - Even better, once you have critical injury tables, you can roll right from the character sheet and it will automatically add a randomized critical injury.
+- Wound State penalties automatically apply as mods on the roll (e.g. if the token is Seriously Wounded all actions will automatically have a -2 penalty)
 
 # DEV WIP BUGFIXES
 - 'Basic' skills (the ones that all characters have points in) now display as bold again (no issue to reference)